+++ cmd/contour/servecontext_test.go
@@ -26,6 +26,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/projectcontour/contour/pkg/config"
+	"k8s.io/utils/pointer"
+
 	contour_api_v1alpha1 "github.com/projectcontour/contour/apis/projectcontour/v1alpha1"
 	envoy_v3 "github.com/projectcontour/contour/internal/envoy/v3"
 	"github.com/projectcontour/contour/internal/fixture"
@@ -357,6 +360,160 @@ func TestParseHTTPVersions(t *testing.T) {
 				func(i, j int) bool { return testcase.parseVersions[i] < testcase.parseVersions[j] })
 
 			assert.Equal(t, testcase.parseVersions, vers)
+		})
+	}
+}
+
+func TestConvertServeContext(t *testing.T) {
+
+	defaultContext := newServeContext()
+	defaultContext.ServerConfig = ServerConfig{
+		xdsAddr:     "127.0.0.1",
+		xdsPort:     8001,
+		caFile:      "/certs/ca.crt",
+		contourCert: "/certs/cert.crt",
+		contourKey:  "/certs/cert.key",
+	}
+
+	defaultContext.ingressClassName = "coolclass"
+	defaultContext.Config.IngressStatusAddress = "1.2.3.4"
+	defaultContext.Config.GatewayConfig = &config.GatewayParameters{
+		ControllerName: "projectcontour.io/projectcontour/contour",
+	}
+	defaultContext.Config.TLS.ClientCertificate = config.NamespacedName{
+		Name:      "cert",
+		Namespace: "secretplace",
+	}
+
+	cases := map[string]struct {
+		serveContext  *serveContext
+		contourConfig contour_api_v1alpha1.ContourConfigurationSpec
+	}{
+		"default ServeContext": {
+			serveContext: defaultContext,
+			contourConfig: contour_api_v1alpha1.ContourConfigurationSpec{
+				XDSServer: contour_api_v1alpha1.XDSServerConfig{
+					Type:    contour_api_v1alpha1.ContourServerType,
+					Address: "127.0.0.1",
+					Port:    8001,
+					TLS: &contour_api_v1alpha1.TLS{
+						CAFile:   "/certs/ca.crt",
+						CertFile: "/certs/cert.crt",
+						KeyFile:  "/certs/cert.key",
+						Insecure: false,
+					},
+				},
+				Ingress: &contour_api_v1alpha1.IngressConfig{
+					ClassName:     pointer.StringPtr("coolclass"),
+					StatusAddress: pointer.StringPtr("1.2.3.4"),
+				},
+				Debug: contour_api_v1alpha1.DebugConfig{
+					Address:                 "127.0.0.1",
+					Port:                    6060,
+					DebugLogLevel:           contour_api_v1alpha1.InfoLog,
+					KubernetesDebugLogLevel: 0,
+				},
+				Health: contour_api_v1alpha1.HealthConfig{
+					Address: "0.0.0.0",
+					Port:    8000,
+				},
+				Envoy: contour_api_v1alpha1.EnvoyConfig{
+					Service: contour_api_v1alpha1.NamespacedName{
+						Name:      "envoy",
+						Namespace: "projectcontour",
+					},
+					HTTPListener: contour_api_v1alpha1.EnvoyListener{
+						Address:   "0.0.0.0",
+						Port:      8080,
+						AccessLog: "/dev/stdout",
+					},
+					HTTPSListener: contour_api_v1alpha1.EnvoyListener{
+						Address:   "0.0.0.0",
+						Port:      8443,
+						AccessLog: "/dev/stdout",
+					},
+					Metrics: contour_api_v1alpha1.MetricsConfig{
+						Address: "0.0.0.0",
+						Port:    8002,
+					},
+					ClientCertificate: &contour_api_v1alpha1.NamespacedName{
+						Name:      "cert",
+						Namespace: "secretplace",
+					},
+					Logging: contour_api_v1alpha1.EnvoyLogging{
+						AccessLogFormat:       contour_api_v1alpha1.EnvoyAccessLog,
+						AccessLogFormatString: nil,
+						AccessLogFields: contour_api_v1alpha1.AccessLogFields([]string{
+							"@timestamp",
+							"authority",
+							"bytes_received",
+							"bytes_sent",
+							"downstream_local_address",
+							"downstream_remote_address",
+							"duration",
+							"method",
+							"path",
+							"protocol",
+							"request_id",
+							"requested_server_name",
+							"response_code",
+							"response_flags",
+							"uber_trace_id",
+							"upstream_cluster",
+							"upstream_host",
+							"upstream_local_address",
+							"upstream_service_time",
+							"user_agent",
+							"x_forwarded_for",
+						}),
+					},
+					DefaultHTTPVersions: nil,
+					Timeouts: &contour_api_v1alpha1.TimeoutParameters{
+						ConnectionIdleTimeout: pointer.StringPtr("60s"),
+					},
+					Cluster: contour_api_v1alpha1.ClusterParameters{
+						DNSLookupFamily: contour_api_v1alpha1.AutoClusterDNSFamily,
+					},
+					Network: contour_api_v1alpha1.NetworkParameters{
+						EnvoyAdminPort: 9001,
+					},
+				},
+				Gateway: &contour_api_v1alpha1.GatewayConfig{
+					ControllerName: "projectcontour.io/projectcontour/contour",
+				},
+				HTTPProxy: contour_api_v1alpha1.HTTPProxyConfig{
+					DisablePermitInsecure: false,
+					FallbackCertificate:   nil,
+				},
+				LeaderElection: contour_api_v1alpha1.LeaderElectionConfig{
+					LeaseDuration: "15s",
+					RenewDeadline: "10s",
+					RetryPeriod:   "2s",
+					Configmap: contour_api_v1alpha1.NamespacedName{
+						Name:      "leader-elect",
+						Namespace: "projectcontour",
+					},
+					DisableLeaderElection: false,
+				},
+				EnableExternalNameService: false,
+				RateLimitService:          nil,
+				Policy: &contour_api_v1alpha1.PolicyConfig{
+					RequestHeadersPolicy:  &contour_api_v1alpha1.HeadersPolicy{},
+					ResponseHeadersPolicy: &contour_api_v1alpha1.HeadersPolicy{},
+					ApplyToIngress:        false,
+				},
+				Metrics: contour_api_v1alpha1.MetricsConfig{
+					Address: "0.0.0.0",
+					Port:    8000,
+				},
+			},
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			converted := tc.serveContext.convertToContourConfigurationSpec()
+			assert.Equal(t, tc.contourConfig, converted)
 		})
 	}
 }
