+++ examples/render/contour-gateway.yaml
@@ -756,6 +756,8 @@ spec:
                 description: Policy specifies default policy applied if not overridden
                   by the user
                 properties:
+                  applyToIngress:
+                    type: boolean
                   requestHeaders:
                     description: RequestHeadersPolicy defines the request headers
                       set/removed on all routes
@@ -1668,6 +1670,8 @@ spec:
                     description: Policy specifies default policy applied if not overridden
                       by the user
                     properties:
+                      applyToIngress:
+                        type: boolean
                       requestHeaders:
                         description: RequestHeadersPolicy defines the request headers
                           set/removed on all routes
