+++ site/content/docs/main/config/api-reference.html
@@ -5409,6 +5409,18 @@ HeadersPolicy
 <p>ResponseHeadersPolicy defines the response headers set/removed on all routes</p>
 </td>
 </tr>
+<tr>
+<td style="white-space:nowrap">
+<code>applyToIngress</code>
+<br>
+<em>
+bool
+</em>
+</td>
+<td>
+<em>(Optional)</em>
+</td>
+</tr>
 </tbody>
 </table>
 <h3 id="projectcontour.io/v1alpha1.RateLimitServiceConfig">RateLimitServiceConfig
